all callbacks that link data to view in updater
