+++ src/components/main_componenets/line_chart.py
@@ -1,8 +1,7 @@
 from dash import Dash, html
-from data.mortgage_data import TotalPaymentRecord
 from .. import ids
 
 
-def render(app: Dash, data: TotalPaymentRecord) -> html.Div:
+def render(app: Dash) -> html.Div:
 
     return html.Div(id=ids.LINE_CHART)
