+++ src/components/modals/edit_mortgage_modal.py
@@ -1,12 +1,10 @@
-from math import floor
 from dash import Dash, html, dcc, Output, Input, State
 import dash_bootstrap_components as dbc
 from .. import ids
-from data.mortgage_data import TotalPaymentRecord
 from datetime import date
 
 
-def render(app: Dash, data: TotalPaymentRecord) -> html.Div:
+def render(app: Dash) -> html.Div:
     @app.callback(
         Output(ids.MORTGAGE_EDIT_MODAL, "is_open"),
         [
@@ -21,49 +19,6 @@ def render(app: Dash, data: TotalPaymentRecord) -> html.Div:
         if n1 or n2:
             return not is_open
         return is_open
-
-    @app.callback(
-        [
-            Output(ids.MORTGAGE_EDIT_MODAL_NAME, "placeholder"),
-            Output(ids.MORTGAGE_EDIT_MODAL_RATE, "placeholder"),
-            Output(ids.MORTGAGE_EDIT_MODAL_START, "initial_visible_month"),
-            Output(ids.MORTGAGE_EDIT_MODAL_START, "placeholder"),
-            Output(ids.MORTGAGE_EDIT_MODAL_FIXED_TERM_YEARS, "placeholder"),
-            Output(ids.MORTGAGE_EDIT_MODAL_FIXED_TERM_MOTHS, "placeholder"),
-            Output(ids.MORTGAGE_EDIT_MODAL_TERM_YEARS, "placeholder"),
-            Output(ids.MORTGAGE_EDIT_MODAL_TERM_MOTHS, "placeholder"),
-            Output(ids.MORTGAGE_EDIT_MODAL_PRINCIPLE, "placeholder"),
-        ],
-        Input(ids.MORTGAGE_AGREEMENT_MODAL_EDIT, "n_clicks"),
-        [
-            State(ids.MORTGAGE_AGREEMENT_MODAL_PREVIOUS, "n_clicks"),
-            State(ids.MORTGAGE_AGREEMENT_MODAL_NEXT, "n_clicks"),
-        ],
-    )
-    def fill_placeholders(
-        n1, prev, next
-    ) -> tuple[str, str, date, str, int, int, int, int, str | None]:
-        """gets the data from the currently displayed mortgage to fill all placeholders"""
-        mortgage = data.mortgage_list[next - prev]
-        name = mortgage.mortgage_name
-        rate = f"{mortgage.interest_rate * 100}%"
-        start_date = mortgage.start_date
-        fixed_years = floor(mortgage.fixed_term / 12)
-        fixed_months = mortgage.fixed_term % 12
-        term_years = floor(mortgage.term / 12)
-        term_months = mortgage.term % 12
-        principle = f"€{mortgage.principle_at_start:,.2f}"
-        return (
-            name,
-            rate,
-            start_date,
-            start_date.strftime("%Y %b"),
-            fixed_years,
-            fixed_months,
-            term_years,
-            term_months,
-            principle,
-        )
 
     return html.Div(
         dbc.Modal(
